refactor: k8s e2e test
